chore: move unit tests in codex-exec into their own files
