+++ codex-rs/exec/src/event_processor_with_jsonl_output_tests.rs
@@ -1,0 +1,54 @@
+use super::*;
+use pretty_assertions::assert_eq;
+use tempfile::tempdir;
+
+#[test]
+fn failed_turn_does_not_overwrite_output_last_message_file() {
+    let tempdir = tempdir().expect("create tempdir");
+    let output_path = tempdir.path().join("last-message.txt");
+    std::fs::write(&output_path, "keep existing contents").expect("seed output file");
+
+    let mut processor = EventProcessorWithJsonOutput::new(Some(output_path.clone()));
+
+    let collected = processor.collect_thread_events(ServerNotification::ItemCompleted(
+        codex_app_server_protocol::ItemCompletedNotification {
+            item: ThreadItem::AgentMessage {
+                id: "msg-1".to_string(),
+                text: "partial answer".to_string(),
+                phase: None,
+                memory_citation: None,
+            },
+            thread_id: "thread-1".to_string(),
+            turn_id: "turn-1".to_string(),
+        },
+    ));
+
+    assert_eq!(collected.status, CodexStatus::Running);
+    assert_eq!(processor.final_message(), Some("partial answer"));
+
+    let status = processor.process_server_notification(ServerNotification::TurnCompleted(
+        codex_app_server_protocol::TurnCompletedNotification {
+            thread_id: "thread-1".to_string(),
+            turn: codex_app_server_protocol::Turn {
+                id: "turn-1".to_string(),
+                items: Vec::new(),
+                status: TurnStatus::Failed,
+                error: Some(codex_app_server_protocol::TurnError {
+                    message: "turn failed".to_string(),
+                    additional_details: None,
+                    codex_error_info: None,
+                }),
+            },
+        },
+    ));
+
+    assert_eq!(status, CodexStatus::InitiateShutdown);
+    assert_eq!(processor.final_message(), None);
+
+    EventProcessor::print_final_output(&mut processor);
+
+    assert_eq!(
+        std::fs::read_to_string(&output_path).expect("read output file"),
+        "keep existing contents"
+    );
+}
